CameraHelpers: implement cameraForBounds
