+++ src/geo/projection/camera_helper.ts
@@ -1,6 +1,10 @@
 import Point from '@mapbox/point-geometry';
 import {IReadonlyTransform, ITransform} from '../transform_interface';
 import {LngLat} from '../lng_lat';
+import {CameraForBoundsOptions} from '../../ui/camera';
+import {PaddingOptions} from '../edge_insets';
+import {LngLatBounds} from '../lng_lat_bounds';
+import {warnOnce} from '../../util/util';
 
 export type MapControlsDeltas = {
     panDelta: Point;
@@ -8,6 +12,15 @@ export type MapControlsDeltas = {
     bearingDelta: number;
     pitchDelta: number;
     around: Point;
+}
+
+/**
+ * @internal
+ */
+export function cameraBoundsWarning() {
+    warnOnce(
+        'Map cannot fit within canvas with the given bounds, padding, and/or offset.'
+    );
 }
 
 /**
@@ -25,4 +38,10 @@ export interface ICameraHelper {
     handleMapControlsPitchBearingZoom(deltas: MapControlsDeltas, tr: ITransform): void;
 
     handleMapControlsPan(deltas: MapControlsDeltas, tr: ITransform, preZoomAroundLoc: LngLat): void;
+
+    cameraForBoxAndBearing(options: CameraForBoundsOptions, padding: PaddingOptions, bounds: LngLatBounds, bearing: number, tr: IReadonlyTransform): {
+        center: LngLat;
+        zoom: number;
+        bearing: number;
+    };
 }
